requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pymongo==3.12.3
python-decouple==3.7
pytz==2022.7
requests==2.28.2
sqlparse==0.2.4
tomli==2.0.1
urllib3==1.26.14
